Condensed user commands into one command
